DD-55-refactor: component/gnb 오류 수정
